@@ -13,29 +13,30 @@ maintainers = [
   {name = "Gabriel Rondeau-Genesse", email = "rondeau-genesse.gabriel@ouranos.ca"}
 ]
 readme = {file = "README.rst", content-type = "text/x-rst"}
-requires-python = ">=3.9.0"
+requires-python = ">=3.10.0"
 keywords = ["xhydro-lstm"]
-license = {file = "LICENSE"}
+license = "Apache-2.0"
+license-files = ["LICENSE"]
 classifiers = [
   "Development Status :: 2 - Pre-Alpha",
   "Intended Audience :: Developers",
-  "License :: OSI Approved :: Apache Software License",
   "Natural Language :: English",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3 :: Only",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
-  # "Programming Language :: Python :: 3.13",
-  "Programming Language :: Python :: Implementation :: CPython"
+  "Programming Language :: Python :: 3.13",
+  # "Programming Language :: Python :: 3.14",
+  "Programming Language :: Python :: Implementation :: CPython",
+  "Typing :: Typed"
 ]
 dynamic = ["description", "version"]
 dependencies = [
   "charset-normalizer >=3.1.0",
-  "netcdf4", # Pinning in RavenPy and xhydro complicate the versioning here.
-  "numpy >=1.24.0",
+  "h5netcdf >=1.5.0",
+  "numpy >=1.25.0",
   "scikit-learn >=1.2.0",
   "tensorflow >=2.10", # CUDA support requires GPU hardware and runners
   "xarray >=2023.11.0",
