@@ -1,12 +1,12 @@
 """Tools to create the datasets to be used in LSTM model training and simulation."""
 
 from __future__ import annotations
-
 import os
 from pathlib import Path
 
 import numpy as np
 import xarray as xr
+
 
 __all__ = [
     "create_dataset_flexible",
