@@ -2,9 +2,9 @@
 """Tests for `xhydro_lstm` package."""
 
 from __future__ import annotations
-
 import pathlib
 from importlib.util import find_spec
+
 
 # import pytest
 
